Release version alpha of the pruning mode support

+++ src/main/kotlin/io/vincenzopalazzo/btcli4j/model/bitcoin/BlockchainInfoBitcoin.kt
@@ -1,14 +1,13 @@
 package io.vincenzopalazzo.btcli4j.model.bitcoin
 
 import com.google.gson.annotations.SerializedName
-import java.math.BigInteger
 
 class BlockchainInfoBitcoin {
     var chain: String? = null
     @SerializedName("initialblockdownload")
     var isDownloading: Boolean? = null
     @SerializedName("headers")
-    var headerCount: BigInteger? = null
+    var headerCount: Long? = null
     @SerializedName("blocks")
-    var blockCount: BigInteger? = null
+    var blockCount: Long? = null
 }
